LPS-54701 Move DDM Form Validator to DDM Validator module
