+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-validator/.project
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <projectDescription>
-	<name>dynamic-data-mapping-form-validator-master</name>
+	<name>dynamic-data-mapping-validator-master</name>
 	<comment></comment>
 	<projects></projects>
 	<buildSpec>
