+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-validator/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay Dynamic Data Mapping Validator
+Bundle-SymbolicName: com.liferay.dynamic.data.mapping.validator
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.dynamic.data.mapping.validator
+Include-Resource: classes
